src/main/java/duke/task/Task.java
package duke.task;

import java.io.Serializable;

public abstract class Task implements Serializable{
    private String description;
    private boolean isDone;

    public Task(String description) {
        this.description = description;
        this.isDone = false;
    }

    public String getDescription() {
        return description;
    }

//    public void setDescription(String description) {
//        this.description = description;
//    }

    public boolean isDone() {
        return isDone;
    }

    public void setDone(boolean isDone) {
        this.isDone = isDone; // Set done status
    }

    private String getStatusIcon() {
        return (isDone ? "X" : " "); // Mark done task with X
    }

    @Override
    public String toString() {
        return "[" + this.getStatusIcon() + "] " + description;
    }
}
